+++ backend/infrastructure-modules/redis-module/src/main/java/com/softeer/throttle/ex/ThrottleException.java
@@ -1,4 +1,4 @@
-package com.softeer.throttle;
+package com.softeer.throttle.ex;
 
 import lombok.Getter;
